refactor(cli): move adot.js logic into run-cluster main waterfall flow
